+++ sdks/python/apache_beam/io/gcp/big_query_query_to_table_it_test.py
@@ -225,7 +225,7 @@ class BigQueryQueryToTableIT(unittest.TestCase):
         'on_success_matcher': all_of(*pipeline_verifiers),
         'kms_key': kms_key,
         'native': True,
-        'experiments': 'use_dataflow_bq_sink',
+        'experiments': 'use_legacy_bq_sink',
     }
     options = self.test_pipeline.get_full_options_as_args(**extra_opts)
     big_query_query_to_table_pipeline.run_bq_pipeline(options)
@@ -304,7 +304,7 @@ class BigQueryQueryToTableIT(unittest.TestCase):
         'native': True,
         'wait_until_finish_duration': WAIT_UNTIL_FINISH_DURATION_MS,
         'on_success_matcher': all_of(*pipeline_verifiers),
-        'experiments': 'use_dataflow_bq_sink',
+        'experiments': 'use_legacy_bq_sink',
     }
     options = self.test_pipeline.get_full_options_as_args(**extra_opts)
     big_query_query_to_table_pipeline.run_bq_pipeline(options)